Change to nosetest unittesting, and ensure jacobi_symbol with denominator 1 is correctly evaluated as 1
