+++ jacobi_symbol.py
@@ -1,6 +1,8 @@
 def jacobi_symbol(a, n):
     """Calculate the Jacobi symbol (a/n)"""
-    if a == 0:
+    if n == 1:
+        return 1
+    elif a == 0:
         return 0
     elif a == 1:
         return 1
